LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookie-store/cookieStore_get_set_across_origins.sub.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

Harness Error (TIMEOUT), message = null

TIMEOUT cookieStore.get() sees cookieStore.set() in cross-origin frame Test timed out
NOTRUN cookieStore.get() in cross-origin frame sees cookieStore.set()
